+++ modules/v1_21_R7_paper/src/test/java/com/nisovin/shopkeepers/compat/v1_21_R7_paper/MappingsVersionTest.java
@@ -20,7 +20,7 @@ public class MappingsVersionTest {
 	public void testMappingsVersion() throws Exception {
 		CompatProviderImpl compatProvider = new CompatProviderImpl();
 		CompatVersion compatVersion = compatProvider.getCompatVersion();
-		String expectedMappingsVersion = compatVersion.getFirstMappingsVersion();
+		String expectedMappingsVersion = compatVersion.getLastMappingsVersion();
 		var actualMappingsVersion = getMinecraftVersion();
 		assertEquals("Unexpected mappings version!",
 				expectedMappingsVersion,