+++ scripts/installSpigotDependencies.sh
@@ -73,7 +73,8 @@ buildSpigotIfMissing 1.21.3 remapped
 buildSpigotIfMissing 1.21.4 remapped
 buildSpigotIfMissing 1.21.5 remapped
 buildSpigotIfMissing 1.21.6 remapped
-buildSpigotIfMissing 1.21.7 remapped
+# Note: 1.21.7 was replaced by 1.21.8 and can no longer be built. But the server is identical to 1.21.8.
+buildSpigotIfMissing 1.21.8 remapped
 buildSpigotIfMissing 1.21.10 remapped
 
 popd